refactor(update-weight): move hf_weight_iterator into training_utils/weight_update/
New home for the unified weight-update engine; modules migrate here
incrementally as the megatron-side update_weight/ decouples.
