+++ miles/backends/megatron_utils/update_weight/update_weight_from_distributed/mixin.py
@@ -7,8 +7,8 @@ import torch
 import torch.distributed as dist
 from tqdm import tqdm
 
-from miles.backends.training_utils.hf_weight_iterator import WeightUpdatePlacement
 from miles.backends.training_utils.parallel import get_parallel_state
+from miles.backends.training_utils.weight_update.hf_weight_iterator import WeightUpdatePlacement
 from miles.utils.distributed_utils import get_gloo_group
 from miles.utils.lora import LORA_ADAPTER_NAME
 from miles.utils.timer import timer
